asset-generation/requirements.txt
google-genai
python-dotenv
Pillow
rembg